Feat. 백로그 정보에 이미지 추가 및 백로그 생성 시 담당자를 지정하여 삽입하는 기능 추가

+++ src/main/java/com/example/peernow360/service/BacklogService.java
@@ -2,7 +2,9 @@ package com.example.peernow360.service;
 
 import com.example.peernow360.dto.BacklogDto;
 import com.example.peernow360.dto.FileDto;
+import com.example.peernow360.dto.UserMemberDto;
 import com.example.peernow360.mappers.IBacklogMapper;
+import com.example.peernow360.mappers.IUserMemberMapper;
 import com.example.peernow360.service.impl.IBacklogService;
 import lombok.RequiredArgsConstructor;
 
@@ -12,6 +14,9 @@ import org.springframework.security.core.userdetails.User;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.StringUtils;
+import org.springframework.web.multipart.MultipartFile;
+
+import java.io.IOException;
 import java.util.*;
 
 @Service
@@ -20,9 +25,11 @@ import java.util.*;
 public class BacklogService implements IBacklogService {
 
     private final IBacklogMapper iBacklogMapper;
+    private final IUserMemberMapper iUserMemberMapper;
+    private final S3Uploader s3Uploader;
 
     @Override
-    public String createNewBacklog(BacklogDto backlogDto, int project_no, String sprint_no, List<FileDto> fileDto) {
+    public String createNewBacklog(BacklogDto backlogDto, int project_no, String sprint_no, MultipartFile[] fileDto) throws IOException {
         log.info("[BacklogService] backlogListInfo()");
 
         Map<String, Object> msgData = new HashMap<>();
@@ -36,10 +43,6 @@ public class BacklogService implements IBacklogService {
 
         log.info("sprint_no: " + Integer.parseInt(sprint_no));
 
-        User user_info = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        String user_id = user_info.getUsername();
-
-        backlogDto.setUser_id(user_id);
         backlogDto.setSprint_no(Integer.parseInt(sprint_no));
         backlogDto.setProject_no(project_no);
 
@@ -50,9 +53,12 @@ public class BacklogService implements IBacklogService {
         if(result > 0) {
             log.info("백로그 생성에 성공하였습니다.");
 
-            if(fileDto != null) {
-                for (FileDto fileName : fileDto) {
-                    msgData.put("name", fileName.getName());
+            if(fileDto != null && fileDto.length > 0) {
+                log.info("이미지가 존재합니다");
+
+                for(MultipartFile fileName : fileDto) {
+                    String storedFileName = s3Uploader.upload(fileName, backlogDto.getUser_id());
+                    msgData.put("name", storedFileName);
                     iBacklogMapper.insertBacklogFile(msgData);
 
                 }
@@ -171,7 +177,7 @@ public class BacklogService implements IBacklogService {
     }
 
     @Override
-    public String backlogUpdateInfo(BacklogDto backlogDto, int no, List<FileDto> fileDto) {
+    public String backlogUpdateInfo(BacklogDto backlogDto, int no, MultipartFile[] fileDto) throws IOException {
         log.info("[BacklogService] backlogUpdateInfo()");
 
         backlogDto.setNo(no);
@@ -181,13 +187,21 @@ public class BacklogService implements IBacklogService {
         if(result > 0) {
             log.info("백로그 수정에 성공하였습니다.");
 
-            int isDel = iBacklogMapper.removeBacklogFile(no);
-            log.info("isDel : 1 -> True / isDel : 0 -> false : " + isDel);
+            if(fileDto != null && fileDto.length > 0) {
+                log.info("이미지가 존재합니다");
 
-            for (FileDto file : fileDto) {
-                file.setBacklog_no(no);
-                iBacklogMapper.insertAndUptBacklogFile(file);
+                Map<String, Object> msgData = new HashMap<>();
+                msgData.put("backlog_no", no);
 
+                int isDel = iBacklogMapper.removeBacklogFile(no);
+                log.info("isDel : 1 -> True / isDel : 0 -> false : " + isDel);
+
+                for (MultipartFile fileName : fileDto) {
+                    String storedFileName = s3Uploader.upload(fileName, backlogDto.getUser_id());
+                    msgData.put("name", storedFileName);
+                    iBacklogMapper.insertBacklogFile(msgData);
+
+                }
             }
 
             return "success";
